Actually support multiple configurations
This is kind of embarassing: since this code changed to support warnings
went in we haven't actually been elaborating our configurations
correctly.  As a result, only one configuration of each step was being
run for each project.

This changes the step mapping code to perform a deep copy, which causes
each target to get a different set of steps.

This has been broken since September.

Fixes: 8edae36 ("Define steps with a dict to configure ShellCommand,
enable warnings")